requirements.txt
pillow
pytesseract
opencv-python
pdf2image
python-dateutil
numpy
Image